fix bug

+++ database/migrations/2026_01_17_204954_add_nama_hrd_to_pengaturan_umum_table.php
@@ -11,9 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('pengaturan_umum', function (Blueprint $table) {
-            $table->string('nama_hrd')->nullable();
-        });
+        if (!Schema::hasColumn('pengaturan_umum', 'nama_hrd')) {
+            Schema::table('pengaturan_umum', function (Blueprint $table) {
+                $table->string('nama_hrd')->nullable();
+            });
+        }
     }
 
     /**
@@ -21,8 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('pengaturan_umum', function (Blueprint $table) {
-            $table->dropColumn('nama_hrd');
-        });
+        if (Schema::hasColumn('pengaturan_umum', 'nama_hrd')) {
+            Schema::table('pengaturan_umum', function (Blueprint $table) {
+                $table->dropColumn('nama_hrd');
+            });
+        }
     }
 };
